Added Homework due for 16/9/2025
This is the quiz about geography for class. I was out sick today (15/9/2025) so I decided to keep the bells and whistles out of if this time, I added a bit to the start but it's a separate function from the actual task. To remove it, delete "quizExtra()" from line 133